backend/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
billiard==4.2.0
blinker==1.5
celery==5.3.5
chardet==5.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cycler==0.12.1
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Cors==4.0.0
Flask-Excel==0.0.7
Flask-JWT-Extended==4.2.3
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.51.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
kombu==5.3.3
lml==0.1.0
MarkupSafe==2.1.2
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
passlib==1.7.4
pdfkit==1.0.0
pillow==10.3.0
prompt-toolkit==3.0.40
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
reportlab==4.1.0
six==1.16.0
SQLAlchemy==2.0.6
texttable==1.7.0
typing_extensions==4.5.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.9
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
